requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
websockets>=16.0,<17.0
SQLAlchemy>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
cryptography>=41.0.0,<43.0.0
psycopg2-binary>=2.9.0,<3.0.0
pymysql>=1.1.0,<2.0.0
